refactor: 홈 화면 legacy component 파일 삭제
