--- src/pages/Home/components/HomeContainer/layout/ScheduleViewMode/index.tsx
@@ -1,3 +1,0 @@
-import ScheduleViewMode from "./ScheduleViewMode.tsx";
-
-export default ScheduleViewMode;
