--- src/pages/Home/components/HomeContainer/view/Calender/boxes/scheduleMarker/ScheduleCount.tsx
@@ -1,15 +1,0 @@
-import { Box } from "@mui/material";
-
-interface ScheduleCountProps {
-  count: number;
-}
-
-function ScheduleCount({ count }: ScheduleCountProps) {
-  return (
-    <Box sx={{ color: "#A9ACB2", fontSize: "12px", textAlign: "center" }}>
-      {count}
-    </Box>
-  );
-}
-
-export default ScheduleCount;
