--- src/pages/Home/components/HomeContainer/view/MonthlyStatement/index.tsx
@@ -1,3 +1,0 @@
-import MonthlyStatement from "./MonthlyStatement.tsx";
-
-export default MonthlyStatement;
