--- src/pages/Home/components/HomeContainer/view/ScheduleList/index.tsx
@@ -1,3 +1,0 @@
-import ScheduleList from "./ScheduleList.tsx";
-
-export default ScheduleList;
